2008-09-28  Sebastien Pouliot  <sebastien@ximian.com>
	* Enumerable.cs: Add missing cast Average on IEnumerable<long?> to
	ensure the result is not truncated.
	[Found using Gendarme's ReviewCastOnIntegerDivisionRule]


svn path=/trunk/mcs/; revision=114330

+++ mcs/class/System.Core/System.Linq/Enumerable.cs
@@ -225,7 +225,7 @@ namespace System.Linq
 		{
 			Check.Source (source);
 
-			return source.AverageNullable<long, long, double> ((a, b) => a + b, (a, b) => a / b);
+			return source.AverageNullable<long, long, double> ((a, b) => a + b, (a, b) => (double) a / b);
 		}
 
 		public static double? Average (this IEnumerable<double?> source)
